requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-binary==122.0.6259.0.0
cinemagoer @ git+https://github.com/cinemagoer/cinemagoer@c33292d692f139e17df76121d7de54dcd67324bb
click==8.1.7
fastapi==0.109.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
streamlit==1.30.0
tenacity==8.2.3
textblob==0.17.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
Unidecode==1.3.8
urllib3==2.1.0
uvicorn==0.26.0
vaderSentiment==3.3.2
validators==0.22.0
watchdog==3.0.0
wsproto==1.2.0
zipp==3.17.0
